itnesspkflushexplode 100% matches
Decompile all remaining itnesspkflushexplode functions:
- it_802AF940 (PK Flash explosion spawn)
- it_802AFA70 (explosion init with camera shake)
- itNesspkflushexplode_UnkMotion0_Anim (scale + hitbox damage)

Fix itPKFlushExplode_ItemVars.xDD8 type from s32 to f32.

+++ src/melee/it/itPKFlash.h
@@ -15,7 +15,7 @@ typedef struct itPKFlush_ItemVars {
 
 typedef struct itPKFlushExplode_ItemVars {
     float xDD4;
-    s32 xDD8;
+    float xDD8;
     HSD_GObj* xDDC;
 } itPKFlushExplode_ItemVars;
